Fix Call Procedure code style.
