+++ hudi-spark-datasource/hudi-spark/src/main/scala/org/apache/spark/sql/hudi/command/procedures/HoodieProcedures.scala
@@ -35,9 +35,9 @@ object HoodieProcedures {
     val mapBuilder: ImmutableMap.Builder[String, Supplier[ProcedureBuilder]] = ImmutableMap.builder()
     mapBuilder.put(RunCompactionProcedure.NAME, RunCompactionProcedure.builder)
     mapBuilder.put(ShowCompactionProcedure.NAME, ShowCompactionProcedure.builder)
-    mapBuilder.put(CreateSavepointsProcedure.NAME, CreateSavepointsProcedure.builder)
-    mapBuilder.put(DeleteSavepointsProcedure.NAME, DeleteSavepointsProcedure.builder)
-    mapBuilder.put(RollbackSavepointsProcedure.NAME, RollbackSavepointsProcedure.builder)
+    mapBuilder.put(CreateSavepointProcedure.NAME, CreateSavepointProcedure.builder)
+    mapBuilder.put(DeleteSavepointProcedure.NAME, DeleteSavepointProcedure.builder)
+    mapBuilder.put(RollbackToSavepointProcedure.NAME, RollbackToSavepointProcedure.builder)
     mapBuilder.put(RollbackToInstantTimeProcedure.NAME, RollbackToInstantTimeProcedure.builder)
     mapBuilder.put(RunClusteringProcedure.NAME, RunClusteringProcedure.builder)
     mapBuilder.put(ShowClusteringProcedure.NAME, ShowClusteringProcedure.builder)
